requirements.txt
click==7.1.1
dnspython==1.16.0
Flask==1.1.1
Flask-API==2.0
Flask-API.yandex==0.6.2.1
Flask-JWT==0.3.2
flask-mongoengine==0.9.5
Flask-PyMongo==2.3.0
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mongoengine==0.19.1
PyJWT==1.4.2
pymongo==3.10.1
pytz==2019.3
six==1.14.0
Werkzeug==1.0.0
WTForms==2.2.1

absl-py==0.9.0
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
gast==0.3.3
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphviz==0.14.1
grpcio==1.31.0
h5py==2.10.0
idna==2.10
joblib==0.16.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.3.1
nltk==3.5
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.0
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
Theano==1.0.5
tqdm==4.48.2
urllib3==1.25.10
wrapt==1.12.1
twilio
transaction